FEAT-#2197: convert docstrings in high level APIs to numpydoc style (#2196)
