+++ .github/workflows/ci.yml
@@ -34,6 +34,23 @@ jobs:
       - run: pip install black
       - run: black --check --diff modin/
 
+  lint-pydocstyle:
+    name: lint (pydocstyle)
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 1
+      - uses: actions/setup-python@v2
+        with:
+          python-version: "3.7.x"
+          architecture: "x64"
+      - run: pip install pydocstyle
+      - run: pydocstyle --convention=numpy --add-ignore=D101,D102,D105 modin/pandas/dataframe.py modin/pandas/series.py modin/pandas/groupby.py modin/pandas/base.py
+      - run: pydocstyle --convention=numpy --add-ignore=D101,D103,D105 modin/pandas/io.py
+      - run: pydocstyle --convention=numpy --add-ignore=D103 modin/pandas/general.py
+      - run: pydocstyle --convention=numpy modin/pandas/plotting.py
+
   lint-flake8:
     name: lint (flake8)
     runs-on: ubuntu-latest
