+++ docs/supported_apis/dataframe_supported.rst
@@ -83,13 +83,17 @@ default to pandas.
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
 | ``copy``                   | `copy`_                   | Y                      |                                                    |
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
-| ``corr``                   | `corr`_                   | Y                      |                                                    |
+| ``corr``                   | `corr`_                   | Y                      | Correlation floating point precision may slightly  |
+|                            |                           |                        | differ from pandas. For now pearson method is      |
+|                            |                           |                        | available only.                                    |
+|                            |                           |                        | For other methods defaults to pandas.              |
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
 | ``corrwith``               | `corrwith`_               | D                      |                                                    |
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
 | ``count``                  | `count`_                  | Y                      |                                                    |
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
-| ``cov``                    | `cov`_                    | Y                      |                                                    |
+| ``cov``                    | `cov`_                    | Y                      | Covariance floating point precision may slightly   |
+|                            |                           |                        | differ from pandas.                                |
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
 | ``cummax``                 | `cummax`_                 | Y                      |                                                    |
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
@@ -213,7 +217,7 @@ default to pandas.
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
 | ``le``                     | `le`_                     | Y                      | See ``add``                                        |
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
-| ``loc``                    | `loc`_                    | Y                      |                                                    |
+| ``loc``                    | `loc`_                    | Y                      | We do not support: boolean array, callable         |
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
 | ``lookup``                 | `lookup`_                 | D                      |                                                    |
 +----------------------------+---------------------------+------------------------+----------------------------------------------------+
