+++ docs/supported_apis/utilities_supported.rst
@@ -21,7 +21,10 @@ default to pandas.
 +---------------------------+---------------------------------+----------------------------------------------------+
 | `pd.unique`_              | Y                               |                                                    |
 +---------------------------+---------------------------------+----------------------------------------------------+
-| ``pd.value_counts``       | Y                               |                                                    |
+| ``pd.value_counts``       | Y                               | The indices of resulting object will be in         |
+|                           |                                 | descending (ascending, if ascending=True) order for|
+|                           |                                 | equal values.                                      |
+|                           |                                 | In pandas indices are located in random order.     |
 +---------------------------+---------------------------------+----------------------------------------------------+
 | `pd.cut`_                 | D                               |                                                    |
 +---------------------------+---------------------------------+----------------------------------------------------+
