+++ modin/pandas/series.py
@@ -11,6 +11,17 @@
 # ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
 
+"""
+Implement Series public API as Pandas does.
+
+Almost all docstrings for public and magic methods should be inherited from Pandas
+for better maintability. So some codes are ignored in pydocstyle check:
+    - D101: missing docstring in class
+    - D102: missing docstring in public method
+    - D105: missing docstring in magic method
+Manually add documentation for methods which are not presented in pandas.
+"""
+
 import numpy as np
 import pandas
 from pandas.core.common import apply_if_callable, is_bool_indexer
@@ -31,15 +42,8 @@ from .base import BasePandasDataset, _ATTRS_NO_LOOKUP
 from .iterator import PartitionIterator
 from .utils import from_pandas, is_scalar
 
-if sys.version_info[0] == 3 and sys.version_info[1] >= 7:
-    # Python >= 3.7
-    from re import Pattern as _pattern_type
-else:
-    # Python <= 3.6
-    from re import _pattern_type
 
-
-@_inherit_docstrings(pandas.Series, excluded=[pandas.Series, pandas.Series.__init__])
+@_inherit_docstrings(pandas.Series, excluded=[pandas.Series.__init__])
 class Series(BasePandasDataset):
     def __init__(
         self,
@@ -54,14 +58,23 @@ class Series(BasePandasDataset):
         """
         One-dimensional ndarray with axis labels (including time series).
 
-        Args:
-            data: Contains data stored in Series.
-            index: Values must be hashable and have the same length as `data`.
-            dtype: Data type for the output Series. If not specified, this will be
-                inferred from `data`.
-            name: The name to give to the Series.
-            copy: Copy input data.
-            query_compiler: A query compiler object to create the Series from.
+        TODO: add types.
+
+        Parameters
+        ----------
+        data:
+            Contains data stored in Series.
+        index:
+            Values must be hashable and have the same length as `data`.
+        dtype:
+            Data type for the output Series. If not specified, this will be
+            inferred from `data`.
+        name:
+            The name to give to the Series.
+        copy:
+            Copy input data.
+        query_compiler: query_compiler
+            A query compiler object to create the Series from.
         """
         if isinstance(data, type(self)):
             query_compiler = data._query_compiler.copy()
@@ -119,54 +132,6 @@ class Series(BasePandasDataset):
     # should be done to columns of parent.
     _parent_axis = 0
 
-    def _reduce_dimension(self, query_compiler):
-        return query_compiler.to_pandas().squeeze()
-
-    def _validate_dtypes_sum_prod_mean(self, axis, numeric_only, ignore_axis=False):
-        return self
-
-    def _validate_dtypes_min_max(self, axis, numeric_only):
-        return self
-
-    def _validate_dtypes(self, numeric_only=False):
-        pass
-
-    def _update_inplace(self, new_query_compiler):
-        super(Series, self)._update_inplace(new_query_compiler=new_query_compiler)
-        # Propagate changes back to parent so that column in dataframe had the same contents
-        if self._parent is not None:
-            if self._parent_axis == 0:
-                self._parent.loc[self.name] = self
-            else:
-                self._parent[self.name] = self
-
-    def _create_or_update_from_compiler(self, new_query_compiler, inplace=False):
-        """Returns or updates a DataFrame given new query_compiler"""
-        assert (
-            isinstance(new_query_compiler, type(self._query_compiler))
-            or type(new_query_compiler) in self._query_compiler.__class__.__bases__
-        ), "Invalid Query Compiler object: {}".format(type(new_query_compiler))
-        if not inplace and new_query_compiler.is_series_like():
-            return Series(query_compiler=new_query_compiler)
-        elif not inplace:
-            # This can happen with things like `reset_index` where we can add columns.
-            from .dataframe import DataFrame
-
-            return DataFrame(query_compiler=new_query_compiler)
-        else:
-            self._update_inplace(new_query_compiler=new_query_compiler)
-
-    def _prepare_inter_op(self, other):
-        if isinstance(other, Series):
-            new_self = self.copy()
-            new_self.name = "__reduced__"
-            new_other = other.copy()
-            new_other.name = "__reduced__"
-        else:
-            new_self = self
-            new_other = other
-        return new_self, new_other
-
     def __add__(self, right):
         return self.add(right)
 
@@ -222,51 +187,7 @@ class Series(BasePandasDataset):
     def __rfloordiv__(self, right):
         return self.rfloordiv(right)
 
-    def _getitem(self, key):
-        key = apply_if_callable(key, self)
-        if isinstance(key, Series) and key.dtype == np.bool:
-            # This ends up being significantly faster than looping through and getting
-            # each item individually.
-            key = key._to_pandas()
-        if is_bool_indexer(key):
-            return self.__constructor__(
-                query_compiler=self._query_compiler.getitem_row_array(
-                    pandas.RangeIndex(len(self.index))[key]
-                )
-            )
-        # TODO: More efficiently handle `tuple` case for `Series.__getitem__`
-        if isinstance(key, tuple):
-            return self._default_to_pandas(pandas.Series.__getitem__, key)
-        else:
-            if not is_list_like(key):
-                reduce_dimension = True
-                key = [key]
-            else:
-                reduce_dimension = False
-            # The check for whether or not `key` is in `keys()` will throw a TypeError
-            # if the object is not hashable. When that happens, we just use the `iloc`.
-            try:
-                if all(k in self.keys() for k in key):
-                    result = self._query_compiler.getitem_row_array(
-                        self.index.get_indexer_for(key)
-                    )
-                else:
-                    result = self._query_compiler.getitem_row_array(key)
-            except TypeError:
-                result = self._query_compiler.getitem_row_array(key)
-        if reduce_dimension:
-            return self._reduce_dimension(result)
-        return self.__constructor__(query_compiler=result)
-
     def __getattr__(self, key):
-        """After regular attribute access, looks up the name in the index
-
-        Args:
-            key (str): Attribute name.
-
-        Returns:
-            The value of the attribute.
-        """
         try:
             return object.__getattribute__(self, key)
         except AttributeError as e:
@@ -365,11 +286,6 @@ class Series(BasePandasDataset):
 
     @property
     def values(self):
-        """Create a NumPy array with the values from this Series.
-
-        Returns:
-            The NumPy representation of this object.
-        """
         return super(Series, self).to_numpy().flatten()
 
     def __xor__(self, other):
@@ -383,32 +299,12 @@ class Series(BasePandasDataset):
         )
 
     def add_prefix(self, prefix):
-        """Add a prefix to each of the column names.
-
-        Returns:
-            A new Series containing the new column names.
-        """
         return Series(query_compiler=self._query_compiler.add_prefix(prefix, axis=0))
 
     def add_suffix(self, suffix):
-        """Add a suffix to each of the column names.
-
-        Returns:
-            A new DataFrame containing the new column names.
-        """
         return Series(query_compiler=self._query_compiler.add_suffix(suffix, axis=0))
 
     def append(self, to_append, ignore_index=False, verify_integrity=False):
-        """Append another DataFrame/list/Series to this one.
-
-        Args:
-            to_append: The object to append to this.
-            ignore_index: Ignore the index on appending.
-            verify_integrity: Verify the integrity of the index on completion.
-
-        Returns:
-            A new DataFrame containing the concatenated values.
-        """
         from .dataframe import DataFrame
 
         bad_type_msg = (
@@ -564,28 +460,6 @@ class Series(BasePandasDataset):
         )
 
     def autocorr(self, lag=1):
-        """
-        Compute the lag-N autocorrelation.
-
-        This method computes the Pearson correlation between
-        the Series and its shifted self.
-
-        Parameters
-        ----------
-        lag : int, default 1
-            Number of lags to apply before performing autocorrelation.
-
-        Returns
-        -------
-        float
-            The Pearson correlation between self and self.shift(lag).
-
-        Notes
-        -----
-        If the Pearson correlation is not well defined return 'NaN'.
-
-        Autocorrelation floating point precision may slightly differ from pandas.
-        """
         return self.corr(self.shift(lag))
 
     def between(self, left, right, inclusive=True):
@@ -605,40 +479,6 @@ class Series(BasePandasDataset):
         keep_shape: bool = False,
         keep_equal: bool = False,
     ):
-        """
-        Compare to another Series and show the differences.
-
-        Parameters
-        ----------
-        other : Series
-            Object to compare with.
-
-        align_axis : {0 or 'index', 1 or 'columns'}, default 1
-            Determine which axis to align the comparison on.
-
-            * 0, or 'index' : Resulting differences are stacked vertically
-                with rows drawn alternately from self and other.
-            * 1, or 'columns' : Resulting differences are aligned horizontally
-                with columns drawn alternately from self and other.
-
-        keep_shape : bool, default False
-            If true, all rows and columns are kept.
-            Otherwise, only the ones with different values are kept.
-
-        keep_equal : bool, default False
-            If true, the result keeps values that are equal.
-            Otherwise, equal values are shown as NaNs.
-
-        Returns
-        -------
-        Series or DataFrame
-            If axis is 0 or 'index' the result will be a Series.
-            The resulting index will be a MultiIndex with 'self' and 'other'
-            stacked alternately at the inner level.
-
-            If axis is 1 or 'columns' the result will be a DataFrame.
-            It will have two columns namely 'self' and 'other'.
-        """
         return self._default_to_pandas(
             pandas.Series.compare,
             other=other,
@@ -648,35 +488,6 @@ class Series(BasePandasDataset):
         )
 
     def corr(self, other, method="pearson", min_periods=None):
-        """
-        Compute correlation with `other` Series, excluding missing values.
-
-        Parameters
-        ----------
-        other : Series
-            Series with which to compute the correlation.
-        method : {'pearson', 'kendall', 'spearman'} or callable
-            Method used to compute correlation:
-
-            - pearson : Standard correlation coefficient
-            - kendall : Kendall Tau correlation coefficient
-            - spearman : Spearman rank correlation
-            - callable: Callable with input two 1d ndarrays and returning a float.
-
-        min_periods : int, optional
-            Minimum number of observations needed to have a valid result.
-
-        Returns
-        -------
-        float
-            Correlation with other.
-
-        Notes
-        -----
-        Correlation floating point precision may slightly differ from pandas.
-
-        For now pearson method is available only. For other methods defaults to pandas.
-        """
         if method == "pearson":
             this, other = self.align(other, join="inner", copy=False)
             this = self.__constructor__(this)
@@ -732,29 +543,6 @@ class Series(BasePandasDataset):
         return super(Series, self).count(level=level)
 
     def cov(self, other, min_periods=None, ddof: Optional[int] = 1):
-        """
-        Compute covariance with Series, excluding missing values.
-
-        Parameters
-        ----------
-        other : Series
-            Series with which to compute the covariance.
-        min_periods : int, optional
-            Minimum number of observations needed to have a valid result.
-        ddof : int, default 1
-            Delta degrees of freedom. The divisor used in calculations is N - ddof,
-            where N represents the number of elements.
-
-        Returns
-        -------
-        float
-            Covariance between Series and other normalized by N-1
-            (unbiased estimator).
-
-        Notes
-        -----
-        Covariance floating point precision may slightly differ from pandas.
-        """
         this, other = self.align(other, join="inner", copy=False)
         this = self.__constructor__(this)
         other = self.__constructor__(other)
@@ -803,38 +591,6 @@ class Series(BasePandasDataset):
         )
 
     def dot(self, other):
-        """
-        Compute the dot product between the Series and the columns of other.
-
-        This method computes the dot product between the Series and another
-        one, or the Series and each columns of a DataFrame, or the Series and
-        each columns of an array.
-
-        It can also be called using `self @ other` in Python >= 3.5.
-
-        Parameters
-        ----------
-        other : Series, DataFrame or array-like
-            The other object to compute the dot product with its columns.
-
-        Returns
-        -------
-        scalar, Series or numpy.ndarray
-            Return the dot product of the Series and other if other is a
-            Series, the Series of the dot product of Series and each rows of
-            other if other is a DataFrame or a numpy.ndarray between the Series
-            and each columns of the numpy array.
-
-        See Also
-        --------
-        DataFrame.dot: Compute the matrix product with the DataFrame.
-        Series.mul: Multiplication of series and other, element-wise.
-
-        Notes
-        -----
-        The Series and other has to share the same index if other is a Series
-        or a DataFrame.
-        """
         if isinstance(other, BasePandasDataset):
             common = self.index.union(other.index)
             if len(common) > len(self.index) or len(common) > len(other.index):
@@ -1062,29 +818,6 @@ class Series(BasePandasDataset):
         )
 
     def median(self, axis=None, skipna=None, level=None, numeric_only=None, **kwargs):
-        """
-        Return the median of the values for the requested axis.
-
-        Parameters
-        ----------
-            axis : {index (0)}
-                Axis for the function to be applied on.
-            skipna : bool, default True
-                Exclude NA/null values when computing the result.
-            level : int or level name, default None
-                If the axis is a MultiIndex (hierarchical), count along a particular level,
-                collapsing into a scalar.
-            numeric_only : bool, default None
-                Include only float, int, boolean columns. If None, will attempt to use everything,
-                then use only numeric data. Not implemented for Series.
-            **kwargs
-                Additional keyword arguments to be passed to the function.
-
-        Returns
-        -------
-        scalar or Series (if level specified)
-            The median of the values for the requested axis
-        """
         axis = self._get_axis_number(axis)
         if numeric_only is not None and not numeric_only:
             self._validate_dtypes(numeric_only=True)
@@ -1142,56 +875,11 @@ class Series(BasePandasDataset):
         return self._default_to_pandas(pandas.Series.nlargest, n=n, keep=keep)
 
     def nsmallest(self, n=5, keep="first"):
-        """
-        Return the smallest `n` elements.
-        Parameters
-        ----------
-        n : int, default 5
-            Return this many ascending sorted values.
-        keep : {'first', 'last', 'all'}, default 'first'
-            When there are duplicate values that cannot all fit in a
-            Series of `n` elements:
-            - ``first`` : return the first `n` occurrences in order
-                of appearance.
-            - ``last`` : return the last `n` occurrences in reverse
-                order of appearance.
-            - ``all`` : keep all occurrences. This can result in a Series of
-                size larger than `n`.
-        Returns
-        -------
-        Series
-            The `n` smallest values in the Series, sorted in increasing order.
-        """
         return Series(query_compiler=self._query_compiler.nsmallest(n=n, keep=keep))
 
     def sem(
         self, axis=None, skipna=None, level=None, ddof=1, numeric_only=None, **kwargs
     ):
-        """
-        Return unbiased standard error of the mean over requested axis.
-
-        Normalized by N-1 by default. This can be changed using the ddof argument
-
-        Parameters
-        ----------
-            axis : {index (0), columns (1)}
-            skipna : bool, default True
-                Exclude NA/null values. If an entire row/column is NA,
-                the result will be NA.
-            level : int or level name, default None
-                If the axis is a MultiIndex (hierarchical), count along a particular level,
-                collapsing into a Series.
-            ddof : int, default 1
-                Delta Degrees of Freedom. The divisor used in calculations is N - ddof,
-                where N represents the number of elements.
-            numeric_only : bool, default None
-                Include only float, int, boolean columns. If None, will attempt to use everything,
-                then use only numeric data. Not implemented for Series.
-
-        Returns
-        -------
-            Scalar or Series (if level specified)
-        """
         axis = self._get_axis_number(axis)
         if numeric_only is not None and not numeric_only:
             self._validate_dtypes(numeric_only=True)
@@ -1218,20 +906,6 @@ class Series(BasePandasDataset):
         )
 
     def slice_shift(self, periods=1, axis=0):
-        """
-        Equivalent to `shift` without copying data.
-        The shifted data will not include the dropped periods and the
-        shifted axis will be smaller than the original.
-        Parameters
-        ----------
-        periods : int
-            Number of periods to move, can be positive or negative.
-        axis : int or str
-            Shift direction.
-        Returns
-        -------
-        shifted : same type as caller
-        """
         if periods == 0:
             return self.copy()
 
@@ -1256,22 +930,6 @@ class Series(BasePandasDataset):
             )
 
     def unstack(self, level=-1, fill_value=None):
-        """
-        Unstack, a.k.a. pivot, Series with MultiIndex to produce DataFrame.
-        The level involved will automatically get sorted.
-
-        Parameters
-        ----------
-        level : int, str, or list of these, default last level
-            Level(s) to unstack, can pass level name.
-        fill_value : scalar value, default None
-            Value to use when replacing NaN values.
-
-        Returns
-        -------
-        DataFrame
-            Unstacked Series.
-        """
         from .dataframe import DataFrame
 
         result = DataFrame(
@@ -1281,29 +939,6 @@ class Series(BasePandasDataset):
         return result.droplevel(0, axis=1) if result.columns.nlevels > 1 else result
 
     def skew(self, axis=None, skipna=None, level=None, numeric_only=None, **kwargs):
-        """
-        Return unbiased skew over requested axis. Normalized by N-1
-
-        Parameters
-        ----------
-            axis : {index (0)}
-                Axis for the function to be applied on.
-            skipna : boolean, default True
-                Exclude NA/null values when computing the result.
-            level : int or level name, default None
-                If the axis is a MultiIndex (hierarchical),
-                count along a particular level, collapsing into a scalar.
-            numeric_only : boolean, default None
-                Include only float, int, boolean columns. If None, will attempt to use everything,
-                then use only numeric data. Not implemented for Series.
-            **kwargs
-                Additional keyword arguments to be passed to the function.
-
-        Returns
-        -------
-        scalar or Series (if level specified)
-            Unbiased skew over requested axis.
-        """
         axis = self._get_axis_number(axis)
         if numeric_only is not None and not numeric_only:
             self._validate_dtypes(numeric_only=True)
@@ -1330,34 +965,6 @@ class Series(BasePandasDataset):
     def std(
         self, axis=None, skipna=None, level=None, ddof=1, numeric_only=None, **kwargs
     ):
-        """
-        Return sample standard deviation over requested axis.
-
-        Normalized by N-1 by default. This can be changed using the ddof argument.
-
-        Parameters
-        ----------
-            axis : {index (0)}
-                The axis to take the std on.
-            skipna : bool, default True
-                Exclude NA/null values. If an entire row/column is NA, the result will be NA.
-            level : int or level name, default None
-                If the axis is a MultiIndex (hierarchical), count along a particular level,
-                collapsing into a scalar.
-            ddof : int, default 1
-                Delta Degrees of Freedom. The divisor used in calculations is N - ddof,
-                where N represents the number of elements.
-            numeric_only : bool, default None
-                Include only float, int, boolean columns. If None, will attempt to use everything,
-                then use only numeric data. Not implemented for Series.
-            **kwargs
-                Additional keyword arguments to be passed to the function.
-
-        Returns
-        -------
-        scalar or Series (if level specified)
-            The sample standard deviation.
-        """
         axis = self._get_axis_number(axis)
         if numeric_only is not None and not numeric_only:
             self._validate_dtypes(numeric_only=True)
@@ -1428,32 +1035,6 @@ class Series(BasePandasDataset):
         min_count=0,
         **kwargs,
     ):
-        """
-        Return the product of the values for the requested axis.
-
-        Parameters
-        ----------
-            axis : {index (0)}
-                Axis for the function to be applied on.
-            skipna : bool, default True
-                Exclude NA/null values when computing the result.
-            level : int or level name, default None
-                If the axis is a MultiIndex (hierarchical), count along a particular level,
-                collapsing into a scalar.
-            numeric_only : bool, default None
-                Include only float, int, boolean columns. If None, will attempt to use everything,
-                then use only numeric data. Not implemented for Series.
-            min_count : int, default 0
-                The required number of valid values to perform the operation.
-                If fewer than min_count non-NA values are present the result will be NA.
-            **kwargs
-                Additional keyword arguments to be passed to the function.
-
-        Returns
-        -------
-        scalar or Series (if level specified)
-            The product of the values for the requested axis.
-        """
         axis = self._get_axis_number(axis)
         new_index = self.columns if axis else self.index
         if min_count > len(new_index):
@@ -1497,19 +1078,6 @@ class Series(BasePandasDataset):
     radd = add
 
     def ravel(self, order="C"):
-        """
-        Returns the flattened containing data as ndarray.
-
-        Parameters
-        ----------
-        order : {'C', 'F', 'A', 'K'}, optional
-
-        Returns
-        ----------
-        numpy.ndarray or ndarray-like
-            Flattened data of the Series.
-
-        """
         data = self._query_compiler.to_numpy().flatten(order=order)
         if isinstance(self.dtype, pandas.CategoricalDtype):
             data = pandas.Categorical(data, dtype=self.dtype)
@@ -1566,27 +1134,6 @@ class Series(BasePandasDataset):
             return result
 
     def repeat(self, repeats, axis=None):
-        """
-        Repeat elements of a Series.
-
-        Returns a new Series where each element of the current Series
-        is repeated consecutively a given number of times.
-
-        Parameters
-        ----------
-        repeats : int or array of ints
-            The number of repetitions for each element. This should be a
-            non-negative integer. Repeating 0 times will return an empty
-            Series.
-        axis : None
-            Must be ``None``. Has no effect but is accepted for compatibility
-            with numpy.
-
-        Returns
-        -------
-        Series
-            Newly created Series with repeated elements.
-        """
         if (isinstance(repeats, int) and repeats == 0) or (
             is_list_like(repeats) and len(repeats) == 1 and repeats[0] == 0
         ):
@@ -1670,44 +1217,6 @@ class Series(BasePandasDataset):
         regex=False,
         method="pad",
     ):
-        """
-        Replace values given in `to_replace` with `value`.
-
-        Values of the Series are replaced with other values dynamically.
-        This differs from updating with .loc or .iloc, which require
-        you to specify a location to update with some value.
-
-        Parameters
-        ----------
-        to_replace : str, regex, list, dict, Series, int, float, or None
-            How to find the values that will be replaced.
-        value : scalar, dict, list, str, regex, default None
-            Value to replace any values matching `to_replace` with.
-            For a DataFrame a dict of values can be used to specify which
-            value to use for each column (columns not in the dict will not be
-            filled). Regular expressions, strings and lists or dicts of such
-            objects are also allowed.
-        inplace : bool, default False
-            If True, in place. Note: this will modify any
-            other views on this object (e.g. a column from a DataFrame).
-            Returns the caller if this is True.
-        limit : int, default None
-            Maximum size gap to forward or backward fill.
-        regex : bool or same types as `to_replace`, default False
-            Whether to interpret `to_replace` and/or `value` as regular
-            expressions. If this is ``True`` then `to_replace` *must* be a
-            string. Alternatively, this could be a regular expression or a
-            list, dict, or array of regular expressions in which case
-            `to_replace` must be ``None``.
-        method : {{'pad', 'ffill', 'bfill', `None`}}
-            The method to use when for replacement, when `to_replace` is a
-            scalar, list or tuple and `value` is ``None``.
-
-        Returns
-        -------
-        Series
-            Object after replacement.
-        """
         inplace = validate_bool_kwarg(inplace, "inplace")
         new_query_compiler = self._query_compiler.replace(
             to_replace=to_replace,
@@ -1720,30 +1229,6 @@ class Series(BasePandasDataset):
         return self._create_or_update_from_compiler(new_query_compiler, inplace)
 
     def searchsorted(self, value, side="left", sorter=None):
-        """
-        Find indices where elements should be inserted to maintain order.
-
-        Find the indices into a sorted Series self such that, if the
-        corresponding elements in value were inserted before the indices,
-        the order of self would be preserved.
-
-        Parameters
-        ----------
-        value: array_like
-            Values to insert into self.
-        side: {"left", "right"}, optional
-            If "left", the index of the first suitable location found is
-            given. If "right", return the last such index. If there is no
-            suitable index, return either 0 or N (where N is the length of self).
-        sorter: 1-D array_like, optional
-            Optional array of integer indices that sort self into ascending order.
-            They are typically the result of np.argsort.
-
-        Returns
-        -------
-        int or array of int
-            A scalar or array of insertion points with the same shape as value.
-        """
         searchsorted_qc = self._query_compiler
         if sorter is not None:
             # `iloc` method works slowly (https://github.com/modin-project/modin/issues/1903),
@@ -1836,32 +1321,6 @@ class Series(BasePandasDataset):
         min_count=0,
         **kwargs,
     ):
-        """
-        Return the sum of the values for the requested axis.
-
-        Parameters
-        ----------
-            axis : {index (0)}
-                Axis for the function to be applied on.
-            skipna : bool, default True
-                Exclude NA/null values when computing the result.
-            level : int or level name, default None
-                If the axis is a MultiIndex (hierarchical), count along a particular level,
-                collapsing into a scalar.
-            numeric_only : bool, default None
-                Include only float, int, boolean columns. If None, will attempt to use everything,
-                then use only numeric data. Not implemented for Series.
-            min_count : int, default 0
-                The required number of valid values to perform the operation.
-                If fewer than min_count non-NA values are present the result will be NA.
-            **kwargs
-                Additional keyword arguments to be passed to the function.
-
-        Returns
-        -------
-        scalar or Series (if level specified)
-            The sum of the values for the requested axis
-        """
         axis = self._get_axis_number(axis)
         new_index = self.columns if axis else self.index
         if min_count > len(new_index):
@@ -1909,32 +1368,6 @@ class Series(BasePandasDataset):
     def take(self, indices, axis=0, is_copy=None, **kwargs):
         return super(Series, self).take(indices, axis=axis, is_copy=is_copy, **kwargs)
 
-    def _to_datetime(self, **kwargs):
-        """
-        Convert `self` to datetime.
-
-        Returns
-        -------
-        datetime
-            Series: Series of datetime64 dtype
-        """
-        return self.__constructor__(
-            query_compiler=self._query_compiler.to_datetime(**kwargs)
-        )
-
-    def _to_numeric(self, **kwargs):
-        """
-        Convert `self` to numeric.
-
-        Returns
-        -------
-        numeric
-            Series: Series of numeric dtype
-        """
-        return self.__constructor__(
-            query_compiler=self._query_compiler.to_numeric(**kwargs)
-        )
-
     def to_dict(self, into=dict):  # pragma: no cover
         return self._default_to_pandas("to_dict", into=into)
 
@@ -1950,28 +1383,6 @@ class Series(BasePandasDataset):
         return self._default_to_pandas(pandas.Series.to_list)
 
     def to_numpy(self, dtype=None, copy=False, na_value=no_default, **kwargs):
-        """
-        A NumPy ndarray representing the values in this Series or Index.
-
-        Parameters
-        ----------
-            dtype : str or numpy.dtype, optional
-                The dtype to pass to numpy.asarray().
-            copy : bool, default False
-                Whether to ensure that the returned value is not a view on another array.
-                Note that copy=False does not ensure that to_numpy() is no-copy.
-                Rather, copy=True ensure that a copy is made, even if not strictly necessary.
-            na_value : Any, optional
-                The value to use for missing values.
-                The default value depends on dtype and the type of the array.
-            **kwargs
-                Additional keywords passed through to
-                the to_numpy method of the underlying array (for extension arrays).
-
-        Returns
-        -------
-            A NumPy array.
-        """
         return (
             super(Series, self)
             .to_numpy(
@@ -2039,28 +1450,9 @@ class Series(BasePandasDataset):
         )
 
     def unique(self):
-        """
-        Return unique values of Series object.
-
-        Uniques are returned in order of appearance. Hash table-based unique,
-        therefore does NOT sort.
-
-        Returns
-        -------
-        ndarray
-            The unique values returned as a NumPy array.
-        """
         return self._query_compiler.unique().to_numpy().squeeze()
 
     def update(self, other):
-        """
-        Modify Series in place using non-NA values from passed
-        Series. Aligns on index.
-
-        Parameters
-        ----------
-        other : Series, or object coercible into Series
-        """
         if not isinstance(other, Series):
             other = Series(other)
         query_compiler = self._query_compiler.series_update(other._query_compiler)
@@ -2069,38 +1461,6 @@ class Series(BasePandasDataset):
     def value_counts(
         self, normalize=False, sort=True, ascending=False, bins=None, dropna=True
     ):
-        """
-        Return a Series containing counts of unique values.
-
-        The resulting object will be in descending order so that the
-        first element is the most frequently-occurring element.
-        Excludes NA values by default.
-
-        Parameters
-        ----------
-        normalize : bool, default False
-            If True then the object returned will contain the relative
-            frequencies of the unique values.
-        sort : bool, default True
-            Sort by frequencies.
-        ascending : bool, default False
-            Sort in ascending order.
-        bins : int, optional
-            Rather than count values, group them into half-open bins,
-            a convenience for ``pd.cut``, only works with numeric data.
-        dropna : bool, default True
-            Don't include counts of NaN.
-
-        Returns
-        -------
-        Series
-
-        Notes
-        -----
-        The indices of resulting object will be in descending
-        (ascending, if ascending=True) order for equal values.
-        It slightly differ from pandas where indices are located in random order.
-        """
         return self.__constructor__(
             query_compiler=self._query_compiler.value_counts(
                 normalize=normalize,
@@ -2114,34 +1474,6 @@ class Series(BasePandasDataset):
     def var(
         self, axis=None, skipna=None, level=None, ddof=1, numeric_only=None, **kwargs
     ):
-        """
-        Return unbiased variance over requested axis.
-
-        Normalized by N-1 by default. This can be changed using the ddof argument
-
-        Parameters
-        ----------
-            axis : {index (0)}
-                The axis to take the variance on.
-            skipna : bool, default True
-                Exclude NA/null values. If an entire row/column is NA, the result will be NA.
-            level : int or level name, default None
-                If the axis is a MultiIndex (hierarchical), count along a particular level,
-                collapsing into a scalar.
-            ddof : int, default 1
-                Delta Degrees of Freedom. The divisor used in calculations is N - ddof,
-                where N represents the number of elements.
-            numeric_only : bool, default None
-                Include only float, int, boolean columns. If None, will attempt to use everything,
-                then use only numeric data. Not implemented for Series.
-            **kwargs
-                Additional keyword arguments to be passed to the function.
-
-        Returns
-        -------
-        scalar or Series (if level specified)
-            The unbiased variance.
-        """
         axis = self._get_axis_number(axis)
         if numeric_only is not None and not numeric_only:
             self._validate_dtypes(numeric_only=True)
@@ -2218,10 +1550,14 @@ class Series(BasePandasDataset):
 
     @property
     def cat(self):
+        from .series_utils import CategoryMethods
+
         return CategoryMethods(self)
 
     @property
     def dt(self):
+        from .series_utils import DatetimeProperties
+
         return DatetimeProperties(self)
 
     @property
@@ -2240,52 +1576,24 @@ class Series(BasePandasDataset):
 
     @property
     def is_monotonic(self):
-        """Return boolean if values in the object are monotonic_increasing.
-
-        Returns
-        -------
-            bool
-        """
         return self._reduce_dimension(self._query_compiler.is_monotonic())
 
     is_monotonic_increasing = is_monotonic
 
     @property
     def is_monotonic_decreasing(self):
-        """Return boolean if values in the object are monotonic_decreasing.
-
-        Returns
-        -------
-            bool
-        """
         return self._reduce_dimension(self._query_compiler.is_monotonic_decreasing())
 
     @property
     def is_unique(self):
-        """Check if Series has no duplicate values.
-
-        Returns
-        -------
-        bool
-            True if there is no duplicates in Series, False otherwise.
-
-        """
         return self.nunique(dropna=False) == len(self)
 
     @property
     def nbytes(self):
-        """
-        Returns the number of bytes in the containing data.
-        """
         return self.memory_usage(index=False)
 
     @property
     def ndim(self):
-        """Get the number of dimensions for this DataFrame.
-
-        Returns:
-            The number of dimensions for this Series.
-        """
         # Series have an invariant that requires they be 1 dimension.
         return 1
 
@@ -2298,609 +1606,227 @@ class Series(BasePandasDataset):
 
     @property
     def str(self):
+        from .series_utils import StringMethods
+
         return StringMethods(self)
 
     def _to_pandas(self):
+        """
+        Implement [METHOD_NAME].
+
+        TODO: Add more details for this docstring template.
+
+        Parameters
+        ----------
+        What arguments does this function have.
+        [
+        PARAMETER_NAME: PARAMETERS TYPES
+            Description.
+        ]
+
+        Returns
+        -------
+        What this returns (if anything)
+        """
         df = self._query_compiler.to_pandas()
         series = df[df.columns[0]]
         if self._query_compiler.columns[0] == "__reduced__":
             series.name = None
         return series
 
+    def _to_datetime(self, **kwargs):
+        """
+        Convert `self` to datetime.
+
+        Returns
+        -------
+        datetime
+            Series: Series of datetime64 dtype
+        """
+        return self.__constructor__(
+            query_compiler=self._query_compiler.to_datetime(**kwargs)
+        )
+
+    def _to_numeric(self, **kwargs):
+        """
+        Convert `self` to numeric.
+
+        Returns
+        -------
+        numeric
+            Series: Series of numeric dtype
+        """
+        return self.__constructor__(
+            query_compiler=self._query_compiler.to_numeric(**kwargs)
+        )
+
+    def _reduce_dimension(self, query_compiler):
+        """
+        Implement [METHOD_NAME].
+
+        TODO: Add more details for this docstring template.
+
+        Parameters
+        ----------
+        What arguments does this function have.
+        [
+        PARAMETER_NAME: PARAMETERS TYPES
+            Description.
+        ]
+
+        Returns
+        -------
+        What this returns (if anything)
+        """
+        return query_compiler.to_pandas().squeeze()
+
+    def _validate_dtypes_sum_prod_mean(self, axis, numeric_only, ignore_axis=False):
+        return self
+
+    def _validate_dtypes_min_max(self, axis, numeric_only):
+        return self
+
+    def _validate_dtypes(self, numeric_only=False):
+        pass
+
+    def _update_inplace(self, new_query_compiler):
+        """
+        Implement [METHOD_NAME].
+
+        TODO: Add more details for this docstring template.
+
+        Parameters
+        ----------
+        What arguments does this function have.
+        [
+        PARAMETER_NAME: PARAMETERS TYPES
+            Description.
+        ]
+
+        Returns
+        -------
+        What this returns (if anything)
+        """
+        super(Series, self)._update_inplace(new_query_compiler=new_query_compiler)
+        # Propagate changes back to parent so that column in dataframe had the same contents
+        if self._parent is not None:
+            if self._parent_axis == 0:
+                self._parent.loc[self.name] = self
+            else:
+                self._parent[self.name] = self
+
+    def _create_or_update_from_compiler(self, new_query_compiler, inplace=False):
+        """
+        Return or update a DataFrame given new query_compiler.
+
+        TODO: add description for parameters.
+
+        Parameters
+        ----------
+        new_query_compiler: query_compiler
+        inplace: bool
+
+        Returns
+        -------
+        Dataframe
+        """
+        assert (
+            isinstance(new_query_compiler, type(self._query_compiler))
+            or type(new_query_compiler) in self._query_compiler.__class__.__bases__
+        ), "Invalid Query Compiler object: {}".format(type(new_query_compiler))
+        if not inplace and new_query_compiler.is_series_like():
+            return Series(query_compiler=new_query_compiler)
+        elif not inplace:
+            # This can happen with things like `reset_index` where we can add columns.
+            from .dataframe import DataFrame
+
+            return DataFrame(query_compiler=new_query_compiler)
+        else:
+            self._update_inplace(new_query_compiler=new_query_compiler)
+
+    def _prepare_inter_op(self, other):
+        """
+        Implement [METHOD_NAME].
+
+        TODO: Add more details for this docstring template.
+
+        Parameters
+        ----------
+        What arguments does this function have.
+        [
+        PARAMETER_NAME: PARAMETERS TYPES
+            Description.
+        ]
+
+        Returns
+        -------
+        What this returns (if anything)
+        """
+        if isinstance(other, Series):
+            new_self = self.copy()
+            new_self.name = "__reduced__"
+            new_other = other.copy()
+            new_other.name = "__reduced__"
+        else:
+            new_self = self
+            new_other = other
+        return new_self, new_other
+
+    def _getitem(self, key):
+        """
+        Implement [METHOD_NAME].
+
+        TODO: Add more details for this docstring template.
+
+        Parameters
+        ----------
+        What arguments does this function have.
+        [
+        PARAMETER_NAME: PARAMETERS TYPES
+            Description.
+        ]
+
+        Returns
+        -------
+        What this returns (if anything)
+        """
+        key = apply_if_callable(key, self)
+        if isinstance(key, Series) and key.dtype == np.bool:
+            # This ends up being significantly faster than looping through and getting
+            # each item individually.
+            key = key._to_pandas()
+        if is_bool_indexer(key):
+            return self.__constructor__(
+                query_compiler=self._query_compiler.getitem_row_array(
+                    pandas.RangeIndex(len(self.index))[key]
+                )
+            )
+        # TODO: More efficiently handle `tuple` case for `Series.__getitem__`
+        if isinstance(key, tuple):
+            return self._default_to_pandas(pandas.Series.__getitem__, key)
+        else:
+            if not is_list_like(key):
+                reduce_dimension = True
+                key = [key]
+            else:
+                reduce_dimension = False
+            # The check for whether or not `key` is in `keys()` will throw a TypeError
+            # if the object is not hashable. When that happens, we just use the `iloc`.
+            try:
+                if all(k in self.keys() for k in key):
+                    result = self._query_compiler.getitem_row_array(
+                        self.index.get_indexer_for(key)
+                    )
+                else:
+                    result = self._query_compiler.getitem_row_array(key)
+            except TypeError:
+                result = self._query_compiler.getitem_row_array(key)
+        if reduce_dimension:
+            return self._reduce_dimension(result)
+        return self.__constructor__(query_compiler=result)
+
 
 if IsExperimental.get():
     from modin.experimental.cloud.meta_magic import make_wrapped_class
 
     make_wrapped_class(Series, "make_series_wrapper")
-
-
-class DatetimeProperties(object):
-    def __init__(self, series):
-        self._series = series
-        self._query_compiler = series._query_compiler
-
-    @property
-    def date(self):
-        return Series(query_compiler=self._query_compiler.dt_date())
-
-    @property
-    def time(self):
-        return Series(query_compiler=self._query_compiler.dt_time())
-
-    @property
-    def timetz(self):
-        return Series(query_compiler=self._query_compiler.dt_timetz())
-
-    @property
-    def year(self):
-        return Series(query_compiler=self._query_compiler.dt_year())
-
-    @property
-    def month(self):
-        return Series(query_compiler=self._query_compiler.dt_month())
-
-    @property
-    def day(self):
-        return Series(query_compiler=self._query_compiler.dt_day())
-
-    @property
-    def hour(self):
-        return Series(query_compiler=self._query_compiler.dt_hour())
-
-    @property
-    def minute(self):
-        return Series(query_compiler=self._query_compiler.dt_minute())
-
-    @property
-    def second(self):
-        return Series(query_compiler=self._query_compiler.dt_second())
-
-    @property
-    def microsecond(self):
-        return Series(query_compiler=self._query_compiler.dt_microsecond())
-
-    @property
-    def nanosecond(self):
-        return Series(query_compiler=self._query_compiler.dt_nanosecond())
-
-    @property
-    def week(self):
-        return Series(query_compiler=self._query_compiler.dt_week())
-
-    @property
-    def weekofyear(self):
-        return Series(query_compiler=self._query_compiler.dt_weekofyear())
-
-    @property
-    def dayofweek(self):
-        return Series(query_compiler=self._query_compiler.dt_dayofweek())
-
-    @property
-    def weekday(self):
-        return Series(query_compiler=self._query_compiler.dt_weekday())
-
-    @property
-    def dayofyear(self):
-        return Series(query_compiler=self._query_compiler.dt_dayofyear())
-
-    @property
-    def quarter(self):
-        return Series(query_compiler=self._query_compiler.dt_quarter())
-
-    @property
-    def is_month_start(self):
-        return Series(query_compiler=self._query_compiler.dt_is_month_start())
-
-    @property
-    def is_month_end(self):
-        return Series(query_compiler=self._query_compiler.dt_is_month_end())
-
-    @property
-    def is_quarter_start(self):
-        return Series(query_compiler=self._query_compiler.dt_is_quarter_start())
-
-    @property
-    def is_quarter_end(self):
-        return Series(query_compiler=self._query_compiler.dt_is_quarter_end())
-
-    @property
-    def is_year_start(self):
-        return Series(query_compiler=self._query_compiler.dt_is_year_start())
-
-    @property
-    def is_year_end(self):
-        return Series(query_compiler=self._query_compiler.dt_is_year_end())
-
-    @property
-    def is_leap_year(self):
-        return Series(query_compiler=self._query_compiler.dt_is_leap_year())
-
-    @property
-    def daysinmonth(self):
-        return Series(query_compiler=self._query_compiler.dt_daysinmonth())
-
-    @property
-    def days_in_month(self):
-        return Series(query_compiler=self._query_compiler.dt_days_in_month())
-
-    @property
-    def tz(self):
-        return self._query_compiler.dt_tz().to_pandas().squeeze()
-
-    @property
-    def freq(self):
-        return self._query_compiler.dt_freq().to_pandas().squeeze()
-
-    def to_period(self, *args, **kwargs):
-        return Series(query_compiler=self._query_compiler.dt_to_period(*args, **kwargs))
-
-    def to_pydatetime(self):
-        return Series(query_compiler=self._query_compiler.dt_to_pydatetime()).to_numpy()
-
-    def tz_localize(self, *args, **kwargs):
-        return Series(
-            query_compiler=self._query_compiler.dt_tz_localize(*args, **kwargs)
-        )
-
-    def tz_convert(self, *args, **kwargs):
-        return Series(
-            query_compiler=self._query_compiler.dt_tz_convert(*args, **kwargs)
-        )
-
-    def normalize(self, *args, **kwargs):
-        return Series(query_compiler=self._query_compiler.dt_normalize(*args, **kwargs))
-
-    def strftime(self, *args, **kwargs):
-        return Series(query_compiler=self._query_compiler.dt_strftime(*args, **kwargs))
-
-    def round(self, *args, **kwargs):
-        return Series(query_compiler=self._query_compiler.dt_round(*args, **kwargs))
-
-    def floor(self, *args, **kwargs):
-        return Series(query_compiler=self._query_compiler.dt_floor(*args, **kwargs))
-
-    def ceil(self, *args, **kwargs):
-        return Series(query_compiler=self._query_compiler.dt_ceil(*args, **kwargs))
-
-    def month_name(self, *args, **kwargs):
-        return Series(
-            query_compiler=self._query_compiler.dt_month_name(*args, **kwargs)
-        )
-
-    def day_name(self, *args, **kwargs):
-        return Series(query_compiler=self._query_compiler.dt_day_name(*args, **kwargs))
-
-    def total_seconds(self, *args, **kwargs):
-        return Series(
-            query_compiler=self._query_compiler.dt_total_seconds(*args, **kwargs)
-        )
-
-    def to_pytimedelta(self):
-        return self._query_compiler.default_to_pandas(
-            lambda df: pandas.Series.dt.to_pytimedelta(df.squeeze(axis=1).dt)
-        )
-
-    @property
-    def seconds(self):
-        return Series(query_compiler=self._query_compiler.dt_seconds())
-
-    @property
-    def days(self):
-        return Series(query_compiler=self._query_compiler.dt_days())
-
-    @property
-    def microseconds(self):
-        return Series(query_compiler=self._query_compiler.dt_microseconds())
-
-    @property
-    def nanoseconds(self):
-        return Series(query_compiler=self._query_compiler.dt_nanoseconds())
-
-    @property
-    def components(self):
-        from .dataframe import DataFrame
-
-        return DataFrame(query_compiler=self._query_compiler.dt_components())
-
-    @property
-    def qyear(self):
-        return Series(query_compiler=self._query_compiler.dt_qyear())
-
-    @property
-    def start_time(self):
-        return Series(query_compiler=self._query_compiler.dt_start_time())
-
-    @property
-    def end_time(self):
-        return Series(query_compiler=self._query_compiler.dt_end_time())
-
-    def to_timestamp(self, *args, **kwargs):
-        return Series(
-            query_compiler=self._query_compiler.dt_to_timestamp(*args, **kwargs)
-        )
-
-
-class StringMethods(object):
-    def __init__(self, series):
-        # Check if dtypes is objects
-
-        self._series = series
-        self._query_compiler = series._query_compiler
-
-    def casefold(self):
-        return self._default_to_pandas(pandas.Series.str.casefold)
-
-    def cat(self, others=None, sep=None, na_rep=None, join=None):
-        if isinstance(others, Series):
-            others = others._to_pandas()
-        return self._default_to_pandas(
-            pandas.Series.str.cat, others=others, sep=sep, na_rep=na_rep, join=join
-        )
-
-    def decode(self, encoding, errors="strict"):
-        return self._default_to_pandas(
-            pandas.Series.str.decode, encoding, errors=errors
-        )
-
-    def split(self, pat=None, n=-1, expand=False):
-        if not pat and pat is not None:
-            raise ValueError("split() requires a non-empty pattern match.")
-
-        if expand:
-            return self._default_to_pandas(
-                pandas.Series.str.split, pat=pat, n=n, expand=expand
-            )
-        else:
-            return Series(
-                query_compiler=self._query_compiler.str_split(
-                    pat=pat, n=n, expand=expand
-                )
-            )
-
-    def rsplit(self, pat=None, n=-1, expand=False):
-        if not pat and pat is not None:
-            raise ValueError("rsplit() requires a non-empty pattern match.")
-
-        if expand:
-            return self._default_to_pandas(
-                pandas.Series.str.rsplit, pat=pat, n=n, expand=expand
-            )
-        else:
-            return Series(
-                query_compiler=self._query_compiler.str_rsplit(
-                    pat=pat, n=n, expand=expand
-                )
-            )
-
-    def get(self, i):
-        return Series(query_compiler=self._query_compiler.str_get(i))
-
-    def join(self, sep):
-        if sep is None:
-            raise AttributeError("'NoneType' object has no attribute 'join'")
-        return Series(query_compiler=self._query_compiler.str_join(sep))
-
-    def get_dummies(self, sep="|"):
-        return self._default_to_pandas(pandas.Series.str.get_dummies, sep=sep)
-
-    def contains(self, pat, case=True, flags=0, na=np.NaN, regex=True):
-        if pat is None and not case:
-            raise AttributeError("'NoneType' object has no attribute 'upper'")
-        return Series(
-            query_compiler=self._query_compiler.str_contains(
-                pat, case=case, flags=flags, na=na, regex=regex
-            )
-        )
-
-    def replace(self, pat, repl, n=-1, case=None, flags=0, regex=True):
-        if not (isinstance(repl, str) or callable(repl)):
-            raise TypeError("repl must be a string or callable")
-        return Series(
-            query_compiler=self._query_compiler.str_replace(
-                pat, repl, n=n, case=case, flags=flags, regex=regex
-            )
-        )
-
-    def repeats(self, repeats):
-        return Series(query_compiler=self._query_compiler.str_repeats(repeats))
-
-    def pad(self, width, side="left", fillchar=" "):
-        if len(fillchar) != 1:
-            raise TypeError("fillchar must be a character, not str")
-        return Series(
-            query_compiler=self._query_compiler.str_pad(
-                width, side=side, fillchar=fillchar
-            )
-        )
-
-    def center(self, width, fillchar=" "):
-        if len(fillchar) != 1:
-            raise TypeError("fillchar must be a character, not str")
-        return Series(
-            query_compiler=self._query_compiler.str_center(width, fillchar=fillchar)
-        )
-
-    def ljust(self, width, fillchar=" "):
-        if len(fillchar) != 1:
-            raise TypeError("fillchar must be a character, not str")
-        return Series(
-            query_compiler=self._query_compiler.str_ljust(width, fillchar=fillchar)
-        )
-
-    def rjust(self, width, fillchar=" "):
-        if len(fillchar) != 1:
-            raise TypeError("fillchar must be a character, not str")
-        return Series(
-            query_compiler=self._query_compiler.str_rjust(width, fillchar=fillchar)
-        )
-
-    def zfill(self, width):
-        return Series(query_compiler=self._query_compiler.str_zfill(width))
-
-    def wrap(self, width, **kwargs):
-        if width <= 0:
-            raise ValueError("invalid width {} (must be > 0)".format(width))
-        return Series(query_compiler=self._query_compiler.str_wrap(width, **kwargs))
-
-    def slice(self, start=None, stop=None, step=None):
-        if step == 0:
-            raise ValueError("slice step cannot be zero")
-        return Series(
-            query_compiler=self._query_compiler.str_slice(
-                start=start, stop=stop, step=step
-            )
-        )
-
-    def slice_replace(self, start=None, stop=None, repl=None):
-        return Series(
-            query_compiler=self._query_compiler.str_slice_replace(
-                start=start, stop=stop, repl=repl
-            )
-        )
-
-    def count(self, pat, flags=0, **kwargs):
-        if not isinstance(pat, (str, _pattern_type)):
-            raise TypeError("first argument must be string or compiled pattern")
-        return Series(
-            query_compiler=self._query_compiler.str_count(pat, flags=flags, **kwargs)
-        )
-
-    def startswith(self, pat, na=np.NaN):
-        return Series(query_compiler=self._query_compiler.str_startswith(pat, na=na))
-
-    def encode(self, encoding, errors="strict"):
-        return self._default_to_pandas(
-            pandas.Series.str.encode, encoding, errors=errors
-        )
-
-    def endswith(self, pat, na=np.NaN):
-        return Series(query_compiler=self._query_compiler.str_endswith(pat, na=na))
-
-    def findall(self, pat, flags=0, **kwargs):
-        if not isinstance(pat, (str, _pattern_type)):
-            raise TypeError("first argument must be string or compiled pattern")
-        return Series(
-            query_compiler=self._query_compiler.str_findall(pat, flags=flags, **kwargs)
-        )
-
-    def match(self, pat, case=True, flags=0, na=np.NaN):
-        if not isinstance(pat, (str, _pattern_type)):
-            raise TypeError("first argument must be string or compiled pattern")
-        return Series(
-            query_compiler=self._query_compiler.str_match(pat, flags=flags, na=na)
-        )
-
-    def extract(self, pat, flags=0, expand=True):
-        return self._default_to_pandas(
-            pandas.Series.str.extract, pat, flags=flags, expand=expand
-        )
-
-    def extractall(self, pat, flags=0):
-        return self._default_to_pandas(pandas.Series.str.extractall, pat, flags=flags)
-
-    def len(self):
-        return Series(query_compiler=self._query_compiler.str_len())
-
-    def strip(self, to_strip=None):
-        return Series(query_compiler=self._query_compiler.str_strip(to_strip=to_strip))
-
-    def rstrip(self, to_strip=None):
-        return Series(query_compiler=self._query_compiler.str_rstrip(to_strip=to_strip))
-
-    def lstrip(self, to_strip=None):
-        return Series(query_compiler=self._query_compiler.str_lstrip(to_strip=to_strip))
-
-    def partition(self, sep=" ", expand=True):
-        if sep is not None and len(sep) == 0:
-            raise ValueError("empty separator")
-
-        if expand:
-            return self._default_to_pandas(
-                pandas.Series.str.partition, sep=sep, expand=expand
-            )
-        else:
-            return Series(
-                query_compiler=self._query_compiler.str_partition(
-                    sep=sep, expand=expand
-                )
-            )
-
-    def repeat(self, repeats):
-        return self._default_to_pandas(pandas.Series.str.repeat, repeats)
-
-    def rpartition(self, sep=" ", expand=True):
-        if sep is not None and len(sep) == 0:
-            raise ValueError("empty separator")
-
-        if expand:
-            return self._default_to_pandas(
-                pandas.Series.str.rpartition, sep=sep, expand=expand
-            )
-        else:
-            return Series(
-                query_compiler=self._query_compiler.str_rpartition(
-                    sep=sep, expand=expand
-                )
-            )
-
-    def lower(self):
-        return Series(query_compiler=self._query_compiler.str_lower())
-
-    def upper(self):
-        return Series(query_compiler=self._query_compiler.str_upper())
-
-    def title(self):
-        return Series(query_compiler=self._query_compiler.str_title())
-
-    def find(self, sub, start=0, end=None):
-        if not isinstance(sub, str):
-            raise TypeError(
-                "expected a string object, not {0}".format(type(sub).__name__)
-            )
-        return Series(
-            query_compiler=self._query_compiler.str_find(sub, start=start, end=end)
-        )
-
-    def rfind(self, sub, start=0, end=None):
-        if not isinstance(sub, str):
-            raise TypeError(
-                "expected a string object, not {0}".format(type(sub).__name__)
-            )
-        return Series(
-            query_compiler=self._query_compiler.str_rfind(sub, start=start, end=end)
-        )
-
-    def index(self, sub, start=0, end=None):
-        if not isinstance(sub, str):
-            raise TypeError(
-                "expected a string object, not {0}".format(type(sub).__name__)
-            )
-        return Series(
-            query_compiler=self._query_compiler.str_index(sub, start=start, end=end)
-        )
-
-    def rindex(self, sub, start=0, end=None):
-        if not isinstance(sub, str):
-            raise TypeError(
-                "expected a string object, not {0}".format(type(sub).__name__)
-            )
-        return Series(
-            query_compiler=self._query_compiler.str_rindex(sub, start=start, end=end)
-        )
-
-    def capitalize(self):
-        return Series(query_compiler=self._query_compiler.str_capitalize())
-
-    def swapcase(self):
-        return Series(query_compiler=self._query_compiler.str_swapcase())
-
-    def normalize(self, form):
-        return Series(query_compiler=self._query_compiler.str_normalize(form))
-
-    def translate(self, table):
-        return Series(query_compiler=self._query_compiler.str_translate(table))
-
-    def isalnum(self):
-        return Series(query_compiler=self._query_compiler.str_isalnum())
-
-    def isalpha(self):
-        return Series(query_compiler=self._query_compiler.str_isalpha())
-
-    def isdigit(self):
-        return Series(query_compiler=self._query_compiler.str_isdigit())
-
-    def isspace(self):
-        return Series(query_compiler=self._query_compiler.str_isspace())
-
-    def islower(self):
-        return Series(query_compiler=self._query_compiler.str_islower())
-
-    def isupper(self):
-        return Series(query_compiler=self._query_compiler.str_isupper())
-
-    def istitle(self):
-        return Series(query_compiler=self._query_compiler.str_istitle())
-
-    def isnumeric(self):
-        return Series(query_compiler=self._query_compiler.str_isnumeric())
-
-    def isdecimal(self):
-        return Series(query_compiler=self._query_compiler.str_isdecimal())
-
-    def _default_to_pandas(self, op, *args, **kwargs):
-        return self._series._default_to_pandas(
-            lambda series: op(series.str, *args, **kwargs)
-        )
-
-
-class CategoryMethods(object):
-    def __init__(self, series):
-        self._series = series
-        self._query_compiler = series._query_compiler
-
-    @property
-    def categories(self):
-        return self._series._default_to_pandas(pandas.Series.cat).categories
-
-    @categories.setter
-    def categories(self, categories):
-        def set_categories(series, categories):
-            series.cat.categories = categories
-
-        self._series._default_to_pandas(set_categories, categories=categories)
-
-    @property
-    def ordered(self):
-        return self._series._default_to_pandas(pandas.Series.cat).ordered
-
-    @property
-    def codes(self):
-        return Series(query_compiler=self._query_compiler.cat_codes())
-
-    def rename_categories(self, new_categories, inplace=False):
-        return self._default_to_pandas(
-            pandas.Series.cat.rename_categories, new_categories, inplace=inplace
-        )
-
-    def reorder_categories(self, new_categories, ordered=None, inplace=False):
-        return self._default_to_pandas(
-            pandas.Series.cat.reorder_categories,
-            new_categories,
-            ordered=ordered,
-            inplace=inplace,
-        )
-
-    def add_categories(self, new_categories, inplace=False):
-        return self._default_to_pandas(
-            pandas.Series.cat.add_categories, new_categories, inplace=inplace
-        )
-
-    def remove_categories(self, removals, inplace=False):
-        return self._default_to_pandas(
-            pandas.Series.cat.remove_categories, removals, inplace=inplace
-        )
-
-    def remove_unused_categories(self, inplace=False):
-        return self._default_to_pandas(
-            pandas.Series.cat.remove_unused_categories, inplace=inplace
-        )
-
-    def set_categories(self, new_categories, ordered=None, rename=False, inplace=False):
-        return self._default_to_pandas(
-            pandas.Series.cat.set_categories,
-            new_categories,
-            ordered=ordered,
-            rename=rename,
-            inplace=inplace,
-        )
-
-    def as_ordered(self, inplace=False):
-        return self._default_to_pandas(pandas.Series.cat.as_ordered, inplace=inplace)
-
-    def as_unordered(self, inplace=False):
-        return self._default_to_pandas(pandas.Series.cat.as_unordered, inplace=inplace)
-
-    def _default_to_pandas(self, op, *args, **kwargs):
-        return self._series._default_to_pandas(
-            lambda series: op(series.cat, *args, **kwargs)
-        )
